put molecule processing functions in more appropriate places
