+++ asapdiscovery-data/asapdiscovery/data/utils.py
@@ -1126,50 +1126,6 @@ def get_compound_id_xtal_dicts(sars_xtals):
     return (compound_to_xtals, xtal_to_compound)
 
 
-def trim_small_chains(input_mol, cutoff_len=10):
-    """
-    Remove short chains from a protein molecule object. The goal is to get rid
-    of any peptide ligands that were mistakenly collected by OESplitMolComplex.
-
-    Parameters
-    ----------
-    input_mol : oechem.OEGraphMol
-        OEGraphMol object containing the protein to trim
-    cutoff_len : int, default=10
-        The cutoff length for peptide chains (chains must have more than this
-        many residues to be included)
-
-    Returns
-    -------
-    oechem.OEGraphMol
-        Trimmed molecule
-    """
-    # Copy the molecule
-    mol_copy = input_mol.CreateCopy()
-
-    # Remove chains from mol_copy that are too short (possibly a better way of
-    #  doing this with OpenEye functions)
-    # Get number of residues per chain
-    chain_len_dict = {}
-    hv = oechem.OEHierView(mol_copy)
-    for chain in hv.GetChains():
-        chain_id = chain.GetChainID()
-        for frag in chain.GetFragments():
-            frag_len = len(list(frag.GetResidues()))
-            try:
-                chain_len_dict[chain_id] += frag_len
-            except KeyError:
-                chain_len_dict[chain_id] = frag_len
-
-    # Remove short chains atom by atom
-    for a in mol_copy.GetAtoms():
-        chain_id = oechem.OEAtomGetResidue(a).GetExtChainID()
-        if (chain_id not in chain_len_dict) or (chain_len_dict[chain_id] <= cutoff_len):
-            mol_copy.DeleteAtom(a)
-
-    return mol_copy
-
-
 def get_ligand_rmsd_openeye(ref: oechem.OEMolBase, mobile: oechem.OEMolBase):
     return oechem.OERMSD(ref, mobile)
