+++ asapdiscovery-docking/asapdiscovery/docking/scripts/calculate_RMSD_of_docked_ligands.py
@@ -20,8 +20,8 @@ from asapdiscovery.data.openeye import (
     load_openeye_sdf,
     load_openeye_sdfs,
     oechem,
-    split_openeye_mol,
 )
+from asapdiscovery.modeling.modeling import split_openeye_mol
 from asapdiscovery.docking.analysis import write_all_rmsds_to_reference
