+++ asapdiscovery-docking/asapdiscovery/docking/scripts/fauxalysis_from_docking.py
@@ -21,7 +21,8 @@ import yaml
 from asapdiscovery.data.openeye import load_openeye_sdf  # noqa: E402
 from asapdiscovery.data.openeye import save_openeye_pdb  # noqa: E402
 from asapdiscovery.data.openeye import save_openeye_sdf  # noqa: E402
-from asapdiscovery.data.openeye import oechem, split_openeye_design_unit
+from asapdiscovery.data.openeye import oechem
+from asapdiscovery.modeling.modeling import split_openeye_design_unit
 from asapdiscovery.docking.analysis import DockingResults  # noqa: E402
