+++ asapdiscovery-docking/asapdiscovery/docking/scripts/make_docked_complexes.py
@@ -8,7 +8,7 @@ import pandas
 from asapdiscovery.data.openeye import load_openeye_pdb  # noqa: E402
 from asapdiscovery.data.openeye import load_openeye_sdf  # noqa: 402
 from asapdiscovery.data.openeye import save_openeye_pdb  # noqa: 402
-from asapdiscovery.data.openeye import split_openeye_mol  # noqa: E402
+from asapdiscovery.modeling.modeling import split_openeye_mol
 from asapdiscovery.data.openeye import combine_protein_ligand, oechem
