+++ asapdiscovery-docking/asapdiscovery/docking/scripts/run_docking_oe.py
@@ -45,8 +45,8 @@ from asapdiscovery.data.openeye import (  # noqa: E402
     oechem,
     save_openeye_pdb,
     save_openeye_sdf,
-    split_openeye_design_unit,
 )
+from asapdiscovery.modeling.modeling import split_openeye_design_unit
 from asapdiscovery.data.schema import ExperimentalCompoundDataUpdate  # noqa: E402
 from asapdiscovery.data.utils import check_filelist_has_elements  # noqa: E402
 from asapdiscovery.docking.docking import run_docking_oe  # noqa: E402
