+++ asapdiscovery-modeling/asapdiscovery/modeling/scripts/prep_fragalysis_dus.py
@@ -13,13 +13,16 @@ import yaml
 from asapdiscovery.data.fragalysis import parse_xtal  # noqa: E402
 from asapdiscovery.data.openeye import save_openeye_pdb  # noqa: E402
 from asapdiscovery.data.openeye import (  # noqa: E402
-    du_to_complex,
     load_openeye_pdb,
     oechem,
 )
 from asapdiscovery.data.utils import edit_pdb_file  # noqa: E402
 from asapdiscovery.data.utils import seqres_to_res_list  # noqa: E402
-from asapdiscovery.modeling.modeling import mutate_residues, prep_receptor
+from asapdiscovery.modeling.modeling import (
+    mutate_residues,
+    prep_receptor,
+    du_to_complex,
+)
 
 
 def check_completed(d):
